perf(vm)!: cut jitless per-instruction, call, and property work
Every execution record now carries its own runtime-budget charge, so
metering no longer re-derives one from the opcode on each dispatched
instruction. Back-edge OSR accounting is gated on the loop-invariant
JIT-installed flag. Callees that capture nothing and bodies without
direct `eval` keep the cell-building and environment-installing paths
out of line.

An ordinary call binds its arguments from one operand-word slice and
pushes its frame directly; only generator entry keeps a separate tail,
so the prepared-frame record is gone. Frame pop resolves the whole
construct/derived/async completion vocabulary from a single cold-pool
probe.

Relational comparison folds Number operands in IEEE-754 directly, and
`Type(v)` primitive classification costs one body type-tag read instead
of three. An own dense array element answers a computed integer index
without first spelling it as a heap string through ToPropertyKey, and a
shape-matched data-slot IC hit no longer re-reads the shape body to
re-prove the slot is not an accessor.

Retired instructions per process on M1: bytecode calls -10.7% at arity
0, -11.2% at arity 4, -11.4% at arity 8; monomorphic method kernel
-9.7%, branch-phi -9.9%, boxed-double property -10.9%, numeric-leaf
-8.5%, dense-array -42.8%.

+++ crates/otter-vm/src/runtime_budget.rs
@@ -275,8 +275,11 @@ impl RuntimeHeapSnapshot {
 }
 
 /// Reduction charge for one executed opcode.
+///
+/// The interpreter never evaluates this at dispatch time: the charge is baked
+/// into every execution record when its owning `CodeBlock` is built.
 #[must_use]
-pub(crate) const fn opcode_reductions(op: Op) -> u64 {
+pub(crate) const fn opcode_reductions(op: Op) -> u8 {
     match op {
         Op::Call
         | Op::CallWithThis
